Commit to karthiga branch with additional files(Feature,Configuration,Schemas,StepDefinition for Morbidity) Moved base and runner file.Kindly review will update accordingly
