+++ src/test/java/StepDefinition/Morbidity_API.java
@@ -1,0 +1,23 @@
+package StepDefinition;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class Morbidity_API {
+	@Given("User want to execute GET operation for Morbidity API")
+	public void user_want_to_execute_get_operation_for_morbidity_api() {
+	   
+	}
+
+	@When("User submit the GET request")
+	public void user_submit_the_get_request() {
+	   
+	}
+
+	@Then("Morbidity data are displayed with {int} OK message")
+	public void morbidity_data_are_displayed_with_ok_message(Integer int1) {
+	   
+	}
+
+}
