+++ src/test/resources/Configuration/config.properties
@@ -1,0 +1,23 @@
+AWS_ACCESS_KEY_ID = james007
+AWS_SECRET_ACCESS_KEY = james007
+
+baseURI = http://127.0.0.1:5000/api
+
+getMorbidityEndPoint= /Morbidity
+getMorbidityNameEndPoint= /Morbidity/MorbidityName=Pre%20Diabetes
+getMorbidityTestIdEndpoint= /Morbidity/MorbidityTestId=PDIA_BG
+postMorbidityEndPoint = /Morbidity/
+deleteMorbidityEndPoint = /Morbidity/MorbidityName=Hypertension&MorbidityTestId=HPT_BP 
+
+getMorbiditySchema=/src/test/resources/Schemas/Morbidity_Schema_API_Get.json
+getMorbididtyByMorbidityName=/src/test/resources/Schemas/Mobidity_Schema_API_Get_MorbidityName.json
+getMorbididtyByMorbidityTestId=/src/test/resources/Schemas/Morbidity_Schema_API_Get_MorbidityTestId.json
+
+postMorbiditySchema=/src/test/resources/Schemas/Morbidity_Schema_API_Post.json
+
+deleteMorbiditySchema=/src/test/resources/Schemas/Morbidity_Schema_API_Delete.json
+
+sheetNameGetMorbidity = GET
+sheetNamePostMorbidity = POST
+sheetNamePutMorbidity = PUT
+sheetNameDeleteMorbidity = DELETE
