+++ src/test/resources/Configuration/cucumber.properties
@@ -1,0 +1,1 @@
+cucumber.publish.quiet=true
